emplemented core functionality on transactions; working on client

+++ backend/src/repo/lotRepository.ts
@@ -2,11 +2,52 @@ import { Database } from "sqlite";
 import { Lot } from "../model/lotModel";
 import { Land } from "../model/landModel";
 
+enum Status {
+  available = "available",
+  reserved = "reserved",
+  sold = "sold",
+}
+
 export class LotRepository {
   constructor(private db: Database) {}
 
   async delete(id: number): Promise<void> {
-    await this.db.run(`DELETE FROM Lot WHERE _id = ?`, [id]);
+    await this.db.exec("BEGIN TRANSACTION");
+
+    try {
+      const lotData = await this.findById(id);
+
+      if (!lotData)
+        throw new Error("Error deleting lot, lot does not exist in DB");
+
+      await this.db.run(`DELETE FROM Lot WHERE _id = ?`, [id]);
+
+      // Case where the lot is marked as Sold(reserve/sold)
+      if (
+        lotData.status === Status.reserved ||
+        lotData.status === Status.sold
+      ) {
+        await this.db.run(
+          `UPDATE Land SET lotsSold = lotsSold - 1 WHERE _id = ?`,
+          [lotData.landId]
+        );
+      } else if (lotData.status === Status.available) {
+        await this.db.run(
+          `UPDATE Land SET available = available - 1 WHERE _id = ?`,
+          [lotData.landId]
+        );
+      }
+
+      await this.db.run(
+        `UPDATE Land SET totalLots = totalLots - 1 WHERE _id = ?`,
+        [lotData.landId]
+      );
+
+      await this.db.exec(`COMMIT`);
+    } catch (error) {
+      await this.db.exec("ROOLBACK");
+      throw error;
+    }
   }
 
   async searchLotsByLandName(payload: {
@@ -93,5 +134,24 @@ export class LotRepository {
     }
 
     return { lots, hasMore };
+  }
+
+  async update(payload: { _id: number; lot: Partial<Lot> }): Promise<void> {
+    const { _id, lot } = payload;
+
+    // filter primary props
+    const fields = Object.keys(lot)
+      .filter((key) => key !== "_id" && key !== "createdAt")
+      .map((key) => `${key} = ?`);
+
+    const values = Object.values(lot);
+
+    // no value to change
+    if (fields.length === 0) return;
+
+    await this.db.run(`UPDATE Lot SET  ${fields.join(", ")} WHERE _id = ?`, [
+      ...values,
+      _id,
+    ]);
   }
 }
